requirements.txt
streamlit
PyPDF2
sentence-transformers
faiss-cpu
openai
python-dotenv